change POSUtlil methods to POS.
