+++ src/main/java/org/spin/pos/service/pos/POS.java
@@ -19,6 +19,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.adempiere.core.domains.models.I_C_Campaign;
+import org.adempiere.core.domains.models.I_C_POS;
+import org.adempiere.exceptions.AdempiereException;
+import org.compiere.model.MPOS;
 import org.compiere.model.MRole;
 import org.compiere.model.Query;
 import org.compiere.util.Env;
@@ -26,6 +29,7 @@ import org.compiere.util.Util;
 import org.spin.backend.grpc.pos.Campaign;
 import org.spin.backend.grpc.pos.ListCampaignsRequest;
 import org.spin.backend.grpc.pos.ListCampaignsResponse;
+import org.spin.base.util.RecordUtil;
 import org.spin.pos.util.POSConvertUtil;
 
 /**
@@ -33,6 +37,56 @@ import org.spin.pos.util.POSConvertUtil;
  * Service Logic for backend of Point Of Sales form
  */
 public class POS {
+
+	/**
+	 * Get POS with identifier
+	 * @param posId
+	 * @param requery
+	 * @return
+	 */
+	public static MPOS validateAndGetPOS(int posId, boolean requery) {
+		if (posId <= 0) {
+			throw new AdempiereException("@FillMandatory@ @C_POS_ID@");
+		}
+		MPOS pos = MPOS.get(Env.getCtx(), posId);
+		if (requery) {
+			pos = new MPOS(Env.getCtx(), posId, null);
+		} else {
+			pos = MPOS.get(Env.getCtx(), posId);
+		}
+		if (pos == null || pos.getC_POS_ID() <= 0) {
+			throw new AdempiereException("@C_POS_ID@ @NotFound@");
+		}
+		return pos;
+	}
+
+	/**
+	 * Get POS with uuid
+	 * @param posId
+	 * @param requery
+	 * @return
+	 */
+	public static MPOS validateAndGetPOS(String posUuid, boolean requery) {
+		if (Util.isEmpty(posUuid, true)) {
+			throw new AdempiereException("@FillMandatory@ @C_POS_ID@");
+		}
+		int posId = RecordUtil.getIdFromUuid(I_C_POS.Table_Name, posUuid, null);
+		return validateAndGetPOS(posId, requery);
+	}
+
+	/**
+	 * Get POS with identifier or uuid
+	 * @param posId
+	 * @param requery
+	 * @return
+	 */
+	public static MPOS validateAndGetPOS(int posId, String posUuid, boolean requery) {
+		if (posId > 0) {
+			return validateAndGetPOS(posId, requery);
+		}
+		return validateAndGetPOS(posUuid, requery);
+	}
+
 	public static ListCampaignsResponse.Builder listCampaigns(ListCampaignsRequest request) {
 		List<Object> filtersList = new ArrayList<>();
